requirements.txt
boto3==1.9.217
botocore==1.12.26
google-api-python-client==1.7.11
google-cloud-storage==1.28.1
paramiko==2.5.0
stratus-api-core>=0.0.3